server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic==2.12.5
openenv-core>=0.1.0
httpx>=0.28.0
openai>=1.40.0
